+++ Amplitude/AMPConstants.h
@@ -23,6 +23,13 @@ extern NSString *const IDENTIFY_EVENT;
 extern NSString *const AMP_OP_ADD;
 extern NSString *const AMP_OP_APPEND;
 extern NSString *const AMP_OP_CLEAR_ALL;
+extern NSString *const AMP_OP_PREPEND;
 extern NSString *const AMP_OP_SET;
 extern NSString *const AMP_OP_SET_ONCE;
 extern NSString *const AMP_OP_UNSET;
+
+extern NSString *const AMP_REVENUE_PRODUCT_ID;
+extern NSString *const AMP_REVENUE_QUANTITY;
+extern NSString *const AMP_REVENUE_PRICE;
+extern NSString *const AMP_REVENUE_REVENUE_TYPE;
+extern NSString *const AMP_REVENUE_RECEIPT;