+++ Amplitude/AMPConstants.m
@@ -5,7 +5,7 @@
 
 NSString *const kAMPLibrary = @"amplitude-ios";
 NSString *const kAMPPlatform = @"iOS";
-NSString *const kAMPVersion = @"3.5.0";
+NSString *const kAMPVersion = @"3.7.0";
 NSString *const kAMPEventLogDomain = @"api.amplitude.com";
 NSString *const kAMPEventLogUrl = @"https://api.amplitude.com/";
 NSString *const kAMPDefaultInstance = @"$default_instance";
@@ -24,6 +24,13 @@ NSString *const IDENTIFY_EVENT = @"$identify";
 NSString *const AMP_OP_ADD = @"$add";
 NSString *const AMP_OP_APPEND = @"$append";
 NSString *const AMP_OP_CLEAR_ALL = @"$clearAll";
+NSString *const AMP_OP_PREPEND = @"$prepend";
 NSString *const AMP_OP_SET = @"$set";
 NSString *const AMP_OP_SET_ONCE = @"$setOnce";
 NSString *const AMP_OP_UNSET = @"$unset";
+
+NSString *const AMP_REVENUE_PRODUCT_ID = @"$productId";
+NSString *const AMP_REVENUE_QUANTITY = @"$quantity";
+NSString *const AMP_REVENUE_PRICE = @"$price";
+NSString *const AMP_REVENUE_REVENUE_TYPE = @"$revenueType";
+NSString *const AMP_REVENUE_RECEIPT = @"$receipt";